// bookshelf_pro/src/components/App/helpers/app-dispatch-type.ts
export enum APP_DISPATCH_TYPE {
    TOGGLE_THEME = 'TOGGLE_THEME',
    ADD_BOOK = 'ADD_BOOK',
    REMOVE_BOOK = 'REMOVE_BOOK',
    TOGGLE_FAVORITE = 'TOGGLE_FAVORITE',
    SET_SEARCH_QUERY = 'SET_SEARCH_QUERY',
    SET_FILTERS = 'SET_FILTERS',
    RESET_FILTER_AND_SEARCH_QUERY = 'RESET_FILTER_AND_SEARCH_QUERY',
    RESET_FAVORITES = 'RESET_FAVORITES',
}